chore: move custom devtools module out of playground
