+++ .nuxtrc
@@ -1,1 +1,2 @@
 typescript.tsConfig.compilerOptions.noUncheckedIndexedAccess=true
+modules[]=../modules/devtools
